Refactor contents widgets and document the tests.
