+++ datashuttle/tui/screens/project_manager.py
@@ -16,8 +16,9 @@ from textual.widgets import (
     TabPane,
 )
 
-from datashuttle.tui.configs import ConfigsContent
-from datashuttle.tui.screens import modal_dialogs, validate
+from datashuttle.tui.screens import modal_dialogs
+from datashuttle.tui.shared.configs_content import ConfigsContent
+from datashuttle.tui.shared.validate_content import ValidateContent
 from datashuttle.tui.tabs import create_folders, logging, transfer
 
 
@@ -74,7 +75,7 @@ class ProjectManagerScreen(Screen):
                     id="tabscreen_transfer_tab",
                 )
             with TabPane("Validate", id="tabscreen_validate_tab"):
-                yield validate.ValidateContent(
+                yield ValidateContent(
                     self, self.interface, id="tabscreen_validate_content"
                 )
             with TabPane("Configs", id="tabscreen_configs_tab"):
