+++ tests/tests_tui/test_tui_validate.py
@@ -2,6 +2,7 @@ import pytest
 import textual
 from tui_base import TuiBase
 
+import datashuttle
 from datashuttle.tui.app import TuiApp
 
 
@@ -11,11 +12,16 @@ class TestTuiValidate(TuiBase):
     async def test_validate_on_project_manager_output(
         self, setup_project_paths
     ):
+        """
+        Check that the validate RichLog is updated as expected.
+        """
         tmp_config_path, tmp_path, project_name = setup_project_paths.values()
 
         app = TuiApp()
         async with app.run_test(size=self.tui_size()) as pilot:
 
+            # Go to the validate tab on project manager, make
+            # some badly formatted files.
             await self.check_and_click_onto_existing_project(
                 pilot, project_name
             )
@@ -44,6 +50,8 @@ class TestTuiValidate(TuiBase):
                 / "ses-badname"
             ).mkdir(parents=True, exist_ok=True)
 
+            # Validate and check validation results are
+            # shown on the Richlog.
             await self.scroll_to_click_pause(
                 pilot, "#validate_validate_button"
             )
@@ -60,17 +68,23 @@ class TestTuiValidate(TuiBase):
     async def test_validate_on_project_manager_kwargs(
         self, setup_project_paths, mocker
     ):
+        """
+        Check options are properly passed through to validate_project
+        from the project manager validate tab (using mocker).
+        """
         tmp_config_path, tmp_path, project_name = setup_project_paths.values()
 
         app = TuiApp()
         async with app.run_test(size=self.tui_size()) as pilot:
+
+            # Set up a project and open the validate tab
             await self.check_and_click_onto_existing_project(
                 pilot, project_name
             )
             await self.switch_tab(pilot, "validate")
 
-            import datashuttle
-
+            # First, check that the default arguments are passed
+            # through to `DataShuttle.validate_project` as expected.
             spy_validate = mocker.spy(
                 datashuttle.utils.validation, "validate_project"
             )
@@ -91,6 +105,9 @@ class TestTuiValidate(TuiBase):
                 "ses": None,
             }
             assert kwargs_["strict_mode"] is False
+
+            # Then, change all arguments and check these are
+            # changed at the level of the called function.
 
             await self.move_select_to_position(
                 pilot, "#validate_top_level_folder_select", position=6
@@ -131,6 +148,7 @@ class TestTuiValidate(TuiBase):
             }
             assert kwargs_["strict_mode"] is False
 
+            # Check the widgets are hidden as expected.
             # Path widgets are not shown for Transfer tab
             for id in [
                 "validate_path_label",
@@ -143,13 +161,17 @@ class TestTuiValidate(TuiBase):
 
     @pytest.mark.asyncio
     async def test_validate_at_path_kwargs(self, setup_project_paths, mocker):
-        """ """
-        # select is not tested, its not criticla and the widget is tested elsewhere.
+        """
+        Test kwargs are properly passed through from the TUI to `quick_validate_project`
+        with mocker. Note that the 'Select' button / directorytree is not tested here,
+        as the screen is tested elsewhere and it's non-critical feature here.
+        """
         tmp_config_path, tmp_path, project_name = setup_project_paths.values()
 
         app = TuiApp()
         async with app.run_test(size=self.tui_size()) as pilot:
 
+            # Open the validation window and input path to project
             project_path = (tmp_path / "local" / project_name).as_posix()
 
             await self.scroll_to_click_pause(
@@ -157,22 +179,23 @@ class TestTuiValidate(TuiBase):
             )
             await self.fill_input(pilot, "#validate_path_input", project_path)
 
-            import datashuttle
-
+            # Spy the function and click 'validate' button
             spy_validate = mocker.spy(
-                datashuttle.tui.screens.validate, "quick_validate_project"
+                datashuttle.tui.shared.validate_content,
+                "quick_validate_project",
             )
 
             await self.scroll_to_click_pause(
                 pilot, "#validate_validate_button"
             )
 
+            # Check args are passed through to function as expected
             args_, kwargs_ = spy_validate.call_args_list[0]
 
             assert args_[0] == project_path
             assert kwargs_["top_level_folder"] == "rawdata"
             assert kwargs_["strict_mode"] is False
 
+            # Check removed widgets, this should be removed because always local
             with pytest.raises(textual.css.query.InvalidQueryFormat):
-                # should be removed because always local
                 pilot.app.query_one("validate_include_central_checkbox")
